analysis/pyproject.toml
[tool.poetry]
name = "vita-app"
version = "0.1.0"
description = ""
authors = ["Son Pham-Ba <son.phamba@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.4.0"
regex = "^2022.10.31"
pandas = "^1.5.3"
openpyxl = "^3.1.0"
levenshtein = "^0.20.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pre-commit = "^3.0.3"
sphinx = "^6.1.3"
sphinx-book-theme = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
preview = true
